PLAT-10113: add script that checks for duplicated users created since last run and merges them
